Update BUGS file and fix some spacings
In particular, the dot space for i and j is increased, and the dot made
bigger. The arrows → &c. now have bigger heads, and I've dicked with the
hinting to try to make them line up with centre dot · .
I've also removed the hints from the shade characters and fixed some
self-intersecting glyphs and overlapping hints.